+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Get meets by date (from @justjasongreen)
+
 ## [1.0.0a0] - 2016-07-21
 ### Added
 - Set up project (from @justjasongreen)
 
+[Unreleased]: https://github.com/justjasongreen/racing_data/compare/1.0.0a0...HEAD
 [1.0.0a0]: https://github.com/justjasongreen/racing_data/tree/1.0.0a0